pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pneumonia-detector"
version = "0.1.0"
description = "CNN-based pneumonia detection from chest X-ray images."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Mobin Yousefi", email = "example@example.com" }
]
keywords = ["deep-learning", "pneumonia", "xray", "cnn", "pytorch"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "scikit-learn>=1.4.0",
  "matplotlib>=3.8.0",
  "tqdm>=4.66.0",
  "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "black>=24.0.0",
  "ruff>=0.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B"]
ignore = ["E501"]